Displaying number of comments on the main page.
