+++ app/__init__.py
@@ -297,7 +297,7 @@ def getSeqAnalysisProjects():
     cache_key = "seq-analysis-projects"
     content = get_cached_data(cache_key)
     if not content:
-        seq_analysis_projects_url = LIMS_API_ROOT + "/LimsRest/getRecentDeliveries"
+        seq_analysis_projects_url = "http://localhost:5007/LimsRest/getRecentDeliveries"
         app.logger.info("Sending request to %s" % seq_analysis_projects_url)
         resp = s.get(seq_analysis_projects_url, auth=(USER, PASSW), verify=False) # , timeout=10)
         content = resp.content
@@ -308,7 +308,7 @@ def getSeqAnalysisProjects():
     incomplete_requests_cache_key = "incomplete-requests"
     incomplete_requests_cached = get_cached_data(incomplete_requests_cache_key)
     if not incomplete_requests_cached:
-        request_statuses_url = LIMS_API_ROOT + "/LimsRest/getSequencingRequests?days=30&delivered=false"
+        request_statuses_url = "http://localhost:5007/LimsRest/getSequencingRequests?days=30&delivered=false"
         app.logger.info("Sending request to %s" % request_statuses_url)
         resp = s.get(request_statuses_url, auth=(USER, PASSW), verify=False) # , timeout=10)
         incomplete_requests_cached = resp.content
@@ -335,6 +335,7 @@ def getSeqAnalysisProjects():
         project['run'] = ', '.join(runs)
         project['ordering'] = recentDate
         project['date'] = time.strftime('%Y-%m-%d %H:%M', time.localtime((recentDate/1000)))
+        project['numComments'] = str(get_num_of_comments(project['requestId']))
 
         if isUnreviewed(project):
             projectsToReview.append(project)
@@ -360,7 +361,7 @@ def getRequestProjects():
     cache_key = "request-projects"
     content = get_cached_data(cache_key)
     if not content:
-        req_projects_url = LIMS_API_ROOT + "/LimsRest/getRecentDeliveries?time=2&units=d"
+        req_projects_url = "http://localhost:5007/LimsRest/getRecentDeliveries?time=2&units=d"
         app.logger.info("Sending request to %s" % req_projects_url)
         resp = s.get(req_projects_url, auth=(USER, PASSW), verify=False) # , timeout=10)
         content = resp.content
@@ -375,6 +376,7 @@ def getRequestProjects():
         [ignore, recentDate] = getRecentDateAndRuns(project)
         project['ordering'] = recentDate
         project['date'] = time.strftime('%Y-%m-%d %H:%M', time.localtime((recentDate/1000)))
+        project['numComments'] = str(get_num_of_comments(project['requestId']))
     projects.sort(key=itemgetter('ordering'))
 
     data = {
@@ -421,6 +423,7 @@ def get_recent_runs():
             project = {}
             mod_timestamp = mtime.strftime("%Y-%m-%d %H:%M")
             project['date'] = mod_timestamp
+            project['numComments'] = str(get_num_of_comments(project['requestId']))
             head, tail = os.path.split(eachfile)
             project['path'] = "static/html/FASTQ/" + tail
             project['runName'] = tail
@@ -433,7 +436,7 @@ def get_recent_runs():
 #route for display the JSON
 @app.route('/JSON_<pId>')
 def displayJSON(pId):
-    r = s.get(LIMS_API_ROOT + "/LimsRest/getProjectQc?project="+pId,  auth=(USER, PASSW), verify=False)
+    r = s.get("http://localhost:5007/LimsRest/getProjectQc?project="+pId,  auth=(USER, PASSW), verify=False)
     data = json.loads(r.content)
     return render_template('json.html', **locals())
 
@@ -469,7 +472,7 @@ def change_run_status():
 
     # Update the cache for this project to reflect the new run status
     project_key = '%s%s' % (CACHE_PROJECT_PREFIX, project)
-    get_project_qc_url = LIMS_API_ROOT + "/LimsRest/getProjectQc?project="+project
+    get_project_qc_url =  "http://localhost:5007/LimsRest/getProjectQc?project="+project
     get_and_cache_project_info(get_project_qc_url, project_key)
 
     success = len(failed_requests) == 0
@@ -491,7 +494,7 @@ def should_repool_sample(recipe, status):
 
 def request_qc_status_change(id, qc_status, project, recipe):
     set_qc_payload = { 'record': id, 'status': qc_status, 'project': project, 'recipe': recipe }
-    set_qc_url = LIMS_API_ROOT  + "/LimsRest/setQcStatus"
+    set_qc_url = "http://localhost:5007/LimsRest/setQcStatus"
 
     resp = s.post(set_qc_url, params=set_qc_payload,  auth=(USER, PASSW), verify=False)
     return 'NewStatus' in resp.text and resp.status_code == 200
@@ -499,7 +502,7 @@ def request_qc_status_change(id, qc_status, project, recipe):
 def request_repool(id, recipe, qc_status):
     app.logger.info("Sending repool request for recipe: %s and status: %s" % (recipe, qc_status))
 
-    set_pooled_url = LIMS_API_ROOT  + "/LimsRest/setPooledSampleStatus"
+    set_pooled_url = "http://localhost:5007/LimsRest/setPooledSampleStatus"
     # Status should be that expected by LIMS for SAMPLES, not the QC site
     set_pooled_payload = {'record': id, 'status': "Ready for - Pooling of Sample Libraries for Sequencing"}
     set_pooled_resp = s.post(set_pooled_url, params=set_pooled_payload,  auth=(USER, PASSW), verify=False)
@@ -513,7 +516,7 @@ def postall_qcStatus(qcStatus, pId):
     recordIds = request.args.getlist('recordIds')
     for recordId in recordIds:
         payload = {'record': recordId, 'status': qcStatus}
-        url = LIMS_API_ROOT  + "/LimsRest/setQcStatus"
+        url = "http://localhost:5007/LimsRest/setQcStatus"
         r = s.post(url, params=payload,  auth=(USER, PASSW), verify=False)
     return make_response(r.text, 200, None)
 
@@ -527,7 +530,7 @@ def add_note():
     """
 
     payload = {'request' : request.form['request'], 'user' : 'qc_review', 'igoUser' : 'gabow', 'readMe' : request.form['readMe']}
-    url = LIMS_API_ROOT + "/LimsRest/limsRequest"
+    url = "http://localhost:5007/LimsRest/limsRequest"
     r =  s.post(url, params=payload,  auth=(USER, PASSW), verify=False)
     return redirect(url_for('index'))
 
@@ -537,7 +540,7 @@ def project_qc(pId):
     if pId == 'undefined' or not is_project_id_valid(pId):
         return create_resp(False, 'Invalid pId', {})
 
-    get_project_qc_resp = s.get(LIMS_API_ROOT + "/LimsRest/getProjectQc?project="+pId, auth=(USER, PASSW), verify=False)
+    get_project_qc_resp = s.get( "http://localhost:5007/LimsRest/getProjectQc?project="+pId, auth=(USER, PASSW), verify=False)
     try:
         get_project_qc = json.loads(get_project_qc_resp.content)
     except TypeError:
@@ -581,13 +584,13 @@ def project_info(pId):
     get_project_qc_resp = get_cached_data(project_key)
     if not get_project_qc_resp:
         app.logger.info('request info not cached, calling limsrest endpoint..')
-        get_project_qc_url = LIMS_API_ROOT + "/LimsRest/getProjectQc?project="+pId
+        get_project_qc_url = "http://localhost:5007/LimsRest/getProjectQc?project="+pId
         get_project_qc_resp = get_and_cache_project_info(get_project_qc_url, project_key)
 
     # TODO - picklist constant
     qc_status_label_content = get_cached_data(CACHE_PICKLIST)
     if not qc_status_label_content:
-        qc_status_label_url = LIMS_API_ROOT + "/LimsRest/getPickListValues?list=Sequencing+QC+Status"
+        qc_status_label_url = "http://localhost:5007/LimsRest/getPickListValues?list=Sequencing+QC+Status"
         app.logger.info('Submitting request to %s' % qc_status_label_url)
         qc_status_label_resp = s.get(qc_status_label_url, auth=(USER, PASSW), verify=False)
         qc_status_label_content = qc_status_label_resp.content
@@ -803,7 +806,7 @@ def get_interops_data():
             cache_data(cache_key, run_summary, CACHE_TIME_LONG)
         return render_template('run_summary.html', run_summary=json.loads(run_summary))
 
-    interops_data_url = LIMS_API_ROOT + "/LimsRest/getInterOpsData?runId="+runName
+    interops_data_url = "http://localhost:5007/LimsRest/getInterOpsData?runId="+runName
     app.logger.info("Sending %s" % interops_data_url)
     r = s.get(interops_data_url, auth=(USER, PASSW), verify=False)
     run_summary = r.content
@@ -830,8 +833,6 @@ def insert_comment():
     myclient = pymongo.MongoClient("mongodb://localhost:27017")
     mydb = myclient["run_qc"]
     mycollection = mydb["qcComments"]
-    # app.logger.info("User commenting is: %s", username)
-    # app.logger.info("User's comment is: %s", comment)
     myquery = {"requestId": pId, "comment": comment, "date": datetime.datetime.now(), "createdBy": username}
     x = mycollection.insert_one(myquery)
     app.logger.info("Inserting comment into qcComments collection is completed.")
@@ -843,7 +844,6 @@ def get_comments(pId):
     mydb = myclient["run_qc"]
     mycollection = mydb["qcComments"]
     myquery = { "requestId" : pId }
-    #app.logger.info("get comment my query is: " + myquery)
     app.logger.info("get comment function is called.")
     mydoc = mycollection.find(myquery)
     print([document for document in mydoc])
@@ -856,7 +856,25 @@ def get_comments(pId):
     
     if not mydoc:
         return create_resp(False, 'No cursor', {})
-    return create_resp(True, 'success', dumps(my_list))     
+    return create_resp(True, 'success', dumps(my_list))
+
+def get_num_of_comments(pId):
+    myclient = pymongo.MongoClient("mongodb://localhost:27017")
+    mydb = myclient["run_qc"]
+    mycollection = mydb["qcComments"]
+    myquery = { "requestId" : pId }
+    mydoc = mycollection.find(myquery)
+    print([document for document in mydoc])
+    my_list = []
+    
+    for document in mydb["qcComments"].find({ "requestId" : pId }):
+        print("document's comment is: " + document["comment"])
+        # insert at beginning of list to order by reserve date
+        my_list.insert(0, document)
+    
+    if not mydoc:
+        return 0
+    return len(my_list)          
 
 if __name__ == '__main__':
     app.run()
