+++ static/seq-qc/src/components/routers/project_router.js
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faArrowAltCircleRight } from '@fortawesome/free-solid-svg-icons'
 import config from '../../config.js';
-import { PROJECT_FLAGS,  LIMS_REQUEST_ID, NUMBER_OF_COMMENTS } from "../../resources/constants";
+import { PROJECT_FLAGS,  LIMS_REQUEST_ID } from "../../resources/constants";
 import {getFlagIcon} from "../project-page/components/quality-checks/quality-checks-utils";
 import Tooltip from '@material-ui/core/Tooltip';
 
@@ -57,7 +57,6 @@ const ProjectRouter = ({name, tooltip, projects}) => {
             for(const project of projects){
                 if (project[field]) return true;
             }
-            console.log("returning false for " + field);
             return false;
         });
         return presentFields;
